+++ test/unit/gemm/device/testbed_universal.h
@@ -133,7 +133,7 @@ struct TestbedUniversal {
     else if (dist_kind == cutlass::Distribution::Sequential) {
 
       cutlass::reference::host::BlockFillSequential(
-        view.data(), view.capacity());
+        view.data(), int64_t(view.capacity()));
     }
     else {
       EXPECT_TRUE(false) << "Not implemented";
